requirements.txt
numpy==1.15.4
scipy==1.1.0
spacy==2.0.16
Flask==1.0.2
gensim==3.6.0
Flask-Cors==3.0.7
PyYAML==3.13
